shard 분리
