+++ src/shared/ui/textarea/index.tsx
@@ -1,0 +1,20 @@
+import { HTMLAttributes, forwardRef } from "react"
+
+interface TextareaProps extends HTMLAttributes<HTMLTextAreaElement> {
+  className?: string
+  rows?: number
+  placeholder: string
+  value: string
+}
+
+// 텍스트 영역 컴포넌트
+export const Textarea = forwardRef<HTMLTextAreaElement, TextareaProps>(({ className, ...props }, ref) => {
+  return (
+    <textarea
+      className={`flex min-h-[150px] w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+      ref={ref}
+      {...props}
+    />
+  )
+})
+Textarea.displayName = "Textarea"
